feat: maintain author-post relationships in post operations
Added logic to ensure author-post relationships are updated when creating, updating, or deleting posts. This includes maintaining the posts array within the author's data and handling changes in authorship when a post is updated.

+++ server/service/post-service.js
@@ -5,6 +5,7 @@ import {
     updatePostValidation
 } from "../validation/post-validation.js";
 import {PostRepository} from "../model/dao/repository/post-repository.js";
+import {UserRepository} from "../model/dao/repository/user-repository.js";
 
 export const PostService = {
     addPost: async (data) => {
@@ -14,7 +15,11 @@ export const PostService = {
             throw new Error(`Field '${validation.error.issues[0].path[0]}' : ${validation.error.issues[0].message}`);
         }
 
-        return await PostRepository.create(data);
+        const newPost = await PostRepository.create(data);
+
+        const author = await UserRepository.getById(data.author);
+        author.posts.push(newPost._id);
+        await author.save();
     },
 
     getPosts: async () => {
@@ -48,7 +53,22 @@ export const PostService = {
             throw new Error(`Field '${validation.error.issues[0].path[0]}' : ${validation.error.issues[0].message}`);
         }
 
-        return await PostRepository.updateById(data._id, data);
+        const post = await PostRepository.getById(data._id);
+
+        if (!post) {
+            throw new Error('Post not found');
+        }
+
+        if (data.author && data.author !== post.author) {
+            const oldAuthor = await UserRepository.getById(post.author.id);
+            oldAuthor.posts.pull(post._id);
+            await oldAuthor.save();
+            const newAuthor = await UserRepository.getById(data.author);
+            newAuthor.posts.push(post._id);
+            await newAuthor.save();
+        }
+
+        await PostRepository.updateById(data._id, data);
     },
 
     deletePost: async (data) => {
@@ -57,6 +77,16 @@ export const PostService = {
         if (!validation.success) {
             throw new Error(`Field '${validation.error.issues[0].path[0]}' : ${validation.error.issues[0].message}`);
         }
+
+        const post = await PostRepository.getById(data.id);
+
+        if (!post) {
+            throw new Error('Post not found');
+        }
+
+        const author = await UserRepository.getById(post.author.id);
+        author.posts.pull(post._id);
+        await author.save();
 
         return await PostRepository.deleteById(data.id);
     }
